+++ packages/auto_animated/example/lib/screens/sliver.dart
@@ -3,15 +3,36 @@ import 'package:flutter/widgets.dart';
 import 'package:auto_animated/auto_animated.dart';
 import 'package:auto_animated_example/utils.dart';
 
-class SliverExample extends StatelessWidget {
+class SliverExample extends StatefulWidget {
+  @override
+  _SliverExampleState createState() => _SliverExampleState();
+}
+
+class _SliverExampleState extends State<SliverExample> {
+  ScrollController _scrollController;
+
+  @override
+  void initState() {
+    _scrollController = ScrollController();
+    super.initState();
+  }
+
+  @override
+  void dispose() {
+    _scrollController.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) => Scaffold(
         body: SafeArea(
           child: CustomScrollView(
+            controller: _scrollController,
             slivers: <Widget>[
               SliverPadding(
                 padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                 sliver: AutoAnimatedSliverList(
+                  controller: _scrollController,
                   showItemInterval: Duration(milliseconds: 250),
                   showItemDuration: Duration(milliseconds: 300),
                   itemCount: 4,
@@ -23,6 +44,7 @@ class SliverExample extends StatelessWidget {
               SliverPadding(
                 padding: EdgeInsets.symmetric(horizontal: 16, vertical: 8),
                 sliver: AutoAnimatedSliverGrid(
+                  controller: _scrollController,
                   delay: Duration(milliseconds: 250) * 5,
                   showItemInterval: Duration(milliseconds: 250),
                   showItemDuration: Duration(milliseconds: 300),