+++ packages/auto_animated/lib/src/grid.dart
@@ -4,14 +4,14 @@ import 'package:flutter/widgets.dart';
 import 'helpers/callbacks.dart';
 import 'helpers/utils.dart' as utils;
 
-const Duration _kDuration = Duration(milliseconds: 300);
+const Duration _kDuration = Duration(milliseconds: 250);
 
 class AutoAnimatedGrid extends StatefulWidget {
   const AutoAnimatedGrid({
     @required this.itemBuilder,
     @required this.gridDelegate,
     @required this.itemCount,
-    this.hideWhenGoingBeyond = true,
+    this.reAnimateOnVisibility = false,
     this.delay = Duration.zero,
     this.showItemInterval = _kDuration,
     this.showItemDuration = _kDuration,
@@ -41,13 +41,8 @@ class AutoAnimatedGrid extends StatefulWidget {
 
   /// Hide the element when it approaches the
   /// frame of the screen so that in the future,
-  /// when it falls into the visibility
-  ///  range, the animation can be played again.
-  ///
-  /// The appearance animation will also play when the item
-  /// is redrawn. Redrawing is peculiar for all
-  ///  list \ grid views with builder methods
-  final bool hideWhenGoingBeyond;
+  /// when it falls into the visibility range - reproduce animation again
+  final bool reAnimateOnVisibility;
 
   /// Called, as needed, to build list item widgets.
   ///
@@ -191,7 +186,7 @@ class _AutoAnimatedGridViewState extends State<AutoAnimatedGrid>
       AnimateOnVisibilityChange(
         key: Key('$_keyPrefix.$itemIndex'),
         duration: widget.showItemDuration,
-        hideWhenGoingBeyond: widget.hideWhenGoingBeyond,
+        reAnimateOnVisibility: widget.reAnimateOnVisibility,
         builder: (context, animation) => widget.itemBuilder(
           context,
           itemIndex,
@@ -203,7 +198,6 @@ class _AutoAnimatedGridViewState extends State<AutoAnimatedGrid>
   Widget build(BuildContext context) => AnimateOnVisibilityWrapper(
         delay: widget.delay,
         showItemInterval: widget.showItemInterval,
-        useListStack: true,
         child: GridView.builder(
           itemBuilder: _itemBuilder,
           gridDelegate: widget.gridDelegate,